# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MeetMind API",
    "version": "1.0.0",
    "description": "REST API for MeetMind, an AI meeting intelligence app. Mobile clients authenticate at POST /api/mobile-auth/login and send the returned token as `Authorization: Bearer <token>`. All endpoints marked with bearerAuth require that token."
  },
  "servers": [
    {
      "url": "/",
      "description": "Same-origin"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/meetings": {
      "get": {
        "operationId": "listMeetings",
        "summary": "List the current user's meetings",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50
            },
            "description": "Maximum number of meetings to return."
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Case-insensitive search over title, transcript and executive summary."
          }
        ],
        "responses": {
          "200": {
            "description": "List of meetings",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "meetings"
                  ],
                  "properties": {
                    "meetings": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/MeetingListItem"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createMeeting",
        "summary": "Create a new draft meeting",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "description": "Meeting title. Defaults to 'Untitled Meeting'."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Created meeting",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "meeting"
                  ],
                  "properties": {
                    "meeting": {
                      "$ref": "#/components/schemas/Meeting"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/meetings/{id}": {
      "get": {
        "operationId": "getMeeting",
        "summary": "Get a single meeting with action items, linked documents and notes",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Meeting identifier (cuid)."
          }
        ],
        "responses": {
          "200": {
            "description": "Meeting detail",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "meeting"
                  ],
                  "properties": {
                    "meeting": {
                      "$ref": "#/components/schemas/MeetingDetail"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "updateMeeting",
        "summary": "Update fields on a meeting",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Meeting identifier (cuid)."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "description": "Any subset of writable meeting fields.",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "status": {
                    "type": "string"
                  },
                  "transcript": {
                    "type": "string"
                  },
                  "executiveSummary": {
                    "type": "string"
                  },
                  "duration": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Operation succeeded",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteMeeting",
        "summary": "Delete a meeting",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Meeting identifier (cuid)."
          }
        ],
        "responses": {
          "200": {
            "description": "Operation succeeded",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/generate-pdf": {
      "post": {
        "operationId": "createPdf",
        "summary": "Start an HTML-to-PDF conversion job",
        "description": "Returns a request_id to poll with /api/generate-pdf/status.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "html_content"
                ],
                "properties": {
                  "html_content": {
                    "type": "string"
                  },
                  "css_stylesheet": {
                    "type": "string",
                    "nullable": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "PDF job created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "request_id": {
                      "type": "string"
                    },
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/generate-pdf/status": {
      "post": {
        "operationId": "getPdfStatus",
        "summary": "Poll the status of an HTML-to-PDF job",
        "description": "Returns status PENDING/RUNNING while processing, SUCCESS with pdf_base64 when done, or FAILED with an error.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "request_id"
                ],
                "properties": {
                  "request_id": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "PDF job status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status"
                  ],
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "pdf_base64": {
                      "type": "string"
                    },
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Bearer token obtained from POST /api/mobile-auth/login."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "Meeting": {
        "type": "object",
        "required": [
          "id",
          "title",
          "date",
          "status",
          "userId",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "duration": {
            "type": "integer",
            "nullable": true
          },
          "status": {
            "type": "string",
            "description": "draft, processing, or completed."
          },
          "transcript": {
            "type": "string",
            "nullable": true
          },
          "rawTranscript": {
            "type": "string",
            "nullable": true
          },
          "executiveSummary": {
            "type": "string",
            "nullable": true
          },
          "keyPoints": {
            "type": "string",
            "nullable": true,
            "description": "JSON-encoded string array."
          },
          "decisions": {
            "type": "string",
            "nullable": true,
            "description": "JSON-encoded string array."
          },
          "openQuestions": {
            "type": "string",
            "nullable": true,
            "description": "JSON-encoded string array."
          },
          "followUps": {
            "type": "string",
            "nullable": true,
            "description": "JSON-encoded string array."
          },
          "userId": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "MeetingListItem": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Meeting"
          },
          {
            "type": "object",
            "properties": {
              "_count": {
                "type": "object",
                "properties": {
                  "actionItems": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        ]
      },
      "MeetingDetail": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Meeting"
          },
          {
            "type": "object",
            "properties": {
              "actionItems": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ActionItem"
                }
              },
              "meetingDocs": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/MeetingDocument"
                }
              },
              "notes": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ComposerNote"
                }
              }
            }
          }
        ]
      },
      "ActionItem": {
        "type": "object",
        "required": [
          "id",
          "task",
          "priority",
          "completed",
          "meetingId"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "task": {
            "type": "string"
          },
          "assignee": {
            "type": "string",
            "nullable": true
          },
          "dueDate": {
            "type": "string",
            "nullable": true
          },
          "priority": {
            "type": "string",
            "description": "High, Medium, or Low."
          },
          "completed": {
            "type": "boolean"
          },
          "meetingId": {
            "type": "string"
          }
        }
      },
      "Document": {
        "type": "object",
        "required": [
          "id",
          "title",
          "fileName",
          "contentType",
          "cloudStoragePath",
          "isPublic",
          "userId",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "fileName": {
            "type": "string"
          },
          "contentType": {
            "type": "string"
          },
          "cloudStoragePath": {
            "type": "string"
          },
          "isPublic": {
            "type": "boolean"
          },
          "extractedText": {
            "type": "string",
            "nullable": true
          },
          "category": {
            "type": "string",
            "nullable": true
          },
          "fileSize": {
            "type": "integer",
            "nullable": true
          },
          "userId": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "MeetingDocument": {
        "type": "object",
        "required": [
          "id",
          "meetingId",
          "documentId"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "meetingId": {
            "type": "string"
          },
          "documentId": {
            "type": "string"
          },
          "document": {
            "$ref": "#/components/schemas/Document"
          }
        }
      },
      "ComposerNote": {
        "type": "object",
        "required": [
          "id",
          "title",
          "userId",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string",
            "nullable": true
          },
          "meetingId": {
            "type": "string",
            "nullable": true
          },
          "userId": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    }
  }
}